fixed routing issues between files from the folder restructuring that were causing the app to crash
