+++ march-madness-draft/client/src/pages/Bracket.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Wrapper from "./components/GameBracket/Wrapper";
-import Regions from "./components/GameBracket/Regions";
+import Wrapper from "../components/GameBracket/Wrapper";
+import Regions from "../components/GameBracket/Regions";
 import API from "../utils/API";
 
 function Bracket(){
